Add files via upload
Final project for the class- we had to make a door lock form scratch that displayed if it was locked or unlocked. The inputs were the buttons on the fpga and the output was the 7 segment. You could set the password, reset the password and put in multiple passwords. We had 3 hours and I finished it in its entirety.